Add preferences for persons
This commit adds the option to attach privileges preferred for each
person when it's created or modified, for this purpose some changes
were made in the people and privileges migration files.

Two partial views were added in the families folder:
- addmemberprivileges.blade.php
- editmemberprivileges.blade.php

Controllers and models were modified too in order to support the new
changes for families and privileges.