Add show and edit options for the Person model
This commit adds show and edit options for the Person model and fixes
some issues in the search method for members.

As an addition, some JS was added in order to obtain the family names
and preset it as surname when a new person is added.

The "Users" option in the menu was relocated under the "Administration"
dropdown option.