Add minimum audit mechanism
This commit adds a minimum mechanism to audit who creates or modifies a
record in the DB, for that reason were added created_by and updated_by
fields in each table. Controllers were modified too to support
the new characteristic.

In addition, every field in every form now has placeholders.